Fix bug in putFile
I think there is an error when adding a file using the `putFile` method if
the file already exists in the database, is not in the memory cache and the
file does not exist in the folder. Then in the file `cache_store.dart` in line
65 a new `CacheObject` is created without a relative path. This throws an
exception when the file `cache_manager.dart` in line 161 is written, since the
file is now written to a location where there is already a folder. I think the
bug can be fixed by adding the relative path from the database to the
`CacheObject`.